feat(recipes): add local embree v4.4.1 recipe with ARM build support
