--- conan-local-recipes-arm64/recipes/embree/all/prebuilt/include/embree4/rtcore.h
@@ -1,14 +1,0 @@
-// Copyright 2009-2021 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "rtcore_config.h"
-#include "rtcore_common.h"
-#include "rtcore_device.h"
-#include "rtcore_buffer.h"
-#include "rtcore_ray.h"
-#include "rtcore_geometry.h"
-#include "rtcore_scene.h"
-#include "rtcore_builder.h"
-#include "rtcore_quaternion.h"
