--- conan-local-recipes-arm64/recipes/embree/all/prebuilt/include/embree4/rtcore.isph
@@ -1,16 +1,0 @@
-// Copyright 2009-2021 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-
-#ifndef __RTC_ISPH__
-#define __RTC_ISPH__
-
-#include "rtcore_config.h"
-#include "rtcore_common.isph"
-#include "rtcore_device.isph"
-#include "rtcore_buffer.isph"
-#include "rtcore_ray.isph"
-#include "rtcore_geometry.isph"
-#include "rtcore_scene.isph"
-#include "rtcore_quaternion.isph"
-
-#endif
