--- conan-local-recipes-arm64/recipes/embree/config.yml
@@ -1,3 +1,0 @@
-versions:
-  "4.4.0":
-    folder: "all"
