+++ conan-local-recipes/recipes/embree/all/patches/4.4.1-0001-fix-arm-clang.patch
@@ -1,0 +1,17 @@
+diff --git a/CMakeLists.txt b/CMakeLists.txt
+index c06124c..038a211 100644
+--- a/CMakeLists.txt
++++ b/CMakeLists.txt
+@@ -288,10 +288,10 @@ IF (WIN32)
+   SET_PROPERTY(GLOBAL PROPERTY USE_FOLDERS ON)
+ 
+   IF ((${CMAKE_CXX_COMPILER_ID} MATCHES "IntelLLVM") OR
+-      (${CMAKE_CXX_COMPILER_ID} MATCHES "Clang" AND COMPILER_HAS_SYCL_SUPPORT))
++      (${CMAKE_CXX_COMPILER_ID} MATCHES "Clang" AND COMPILER_HAS_SYCL_SUPPORT AND NOT EMBREE_ARM))
+     MESSAGE("-- DPCPP compiler detected")
+     INCLUDE(dpcpp)
+-  ELSEIF(${CMAKE_GENERATOR_TOOLSET} MATCHES "^LLVM")
++  ELSEIF(${CMAKE_GENERATOR_TOOLSET} MATCHES "^LLVM" OR ${CMAKE_CXX_COMPILER_ID} MATCHES "Clang")
+     set(EMBREE_SYCL_SUPPORT OFF)
+     MESSAGE("-- CLANG compiler detected")
+     INCLUDE(clang)
