+++ conan-local-recipes/recipes/embree/config.yml
@@ -1,0 +1,3 @@
+versions:
+  "4.4.1":
+    folder: "all"
