+++ conanfile.py
@@ -14,7 +14,7 @@ import os
 BOOST_VERSION = "1.88.0"
 DOXYGEN_VERSION = "1.17.0"
 EIGEN_VERSION = "5.0.1"
-EMBREE_VERSION = "4.4.0"
+EMBREE_VERSION = "4.4.1"
 FMT_VERSION = "12.1.0"
 GLFW_VERSION = "3.4"
 IMATH_VERSION = "3.2.2"
