+++ run-conan.sh
@@ -90,40 +90,6 @@ conan remote add mycenter ./conan-center-index --force
 conan remote add mylocal ./conan-local-recipes --force
 conan list -r mylocal
 
-# Add ARM64-specific recipes only on ARM64
-if [[ "$RUNNER_OS" == "Windows" && "$RUNNER_ARCH" == "ARM64" ]]; then
-  conan remote add mylocal-arm64 ./conan-local-recipes-arm64 --force
-  conan list -r mylocal-arm64
-fi
-echo "::endgroup::"
-
-# 2.5 Only create local embree package on ARM64
-if [[ "$RUNNER_OS" == "Windows" && "$RUNNER_ARCH" == "ARM64" ]];then
-  echo "::group::CIBW_BEFORE_BUILD: Build local packages from source"
-
-  # Step A: build onetbb and hwloc with clang profile
-  # so package IDs match what Embree expects
-  conan install \
-    --requires=onetbb/2022.3.0 \
-    --profile:all=$CLANG_PROFILE_ARM64 \
-    --build=missing \
-    --remote=mycenter \
-    --remote=mylocal \
-    --remote=conancenter
-
-  # Step B: build Embree with clang profile
-  # package_id() removes compiler so MSVC consumers find it in cache
-  cd $WORKSPACE/conan-local-recipes-arm64/recipes/embree/all
-  echo "Current directory: $(pwd)"
-  echo "Compiling latest Embree source via Clang-cl..."
-  conan create . \
-    --profile:all=$CLANG_PROFILE_ARM64 \
-    --version=4.4.0 \
-    --build=embree*
-  cd $WORKSPACE
-  echo "::endgroup::"
-fi
-
 if [[ "$RUNNER_OS" == "Linux" ]]; then
   # ispc
   echo "::group::CIBW_BEFORE_BUILD: ispc"
@@ -179,24 +145,13 @@ fi
 echo "::group::CIBW_BEFORE_BUILD: Explain graph"
 # This is only for debugging purpose...
 cd $WORKSPACE
-if [[ "$RUNNER_OS" == "Windows" && "$RUNNER_ARCH" == "ARM64" ]]; then
-  conan graph info $WORKSPACE \
-    --profile:all=$CONAN_PROFILE \
-    --version=$LUXDEPS_VERSION \
-    --remote=mycenter \
-    --remote=mylocal \
-    --remote=mylocal-arm64 \
-    --build=missing \
-    --format=dot
-else
-  conan graph info $WORKSPACE \
-    --profile:all=$CONAN_PROFILE \
-    --version=$LUXDEPS_VERSION \
-    --remote=mycenter \
-    --remote=mylocal \
-    --build=missing \
-    --format=dot
-fi
+conan graph info $WORKSPACE \
+  --profile:all=$CONAN_PROFILE \
+  --version=$LUXDEPS_VERSION \
+  --remote=mycenter \
+  --remote=mylocal \
+  --build=missing \
+  --format=dot
 echo "::endgroup::"
 
 # (Debug) Install particular package, for debugging
@@ -211,24 +166,29 @@ echo "::endgroup::"
 # precompiled binaries and it forces compilation)
 echo "::group::CIBW_BEFORE_BUILD: Create LuxCore Deps"
 cd $WORKSPACE
+
+EMBREE_PROFILE=$CONAN_PROFILE
 if [[ "$RUNNER_OS" == "Windows" && "$RUNNER_ARCH" == "ARM64" ]]; then
-  conan create $WORKSPACE \
-    --profile:all=$CONAN_PROFILE \
-    --version=$LUXDEPS_VERSION \
-    --remote=mycenter \
-    --remote=mylocal \
-    --remote=mylocal-arm64 \
-    --remote=conancenter \
-    --build=missing \
-    --build=!embree*   # ← never rebuild embree, use whatever is in cache
-else
-  conan create $WORKSPACE \
-    --profile:all=$CONAN_PROFILE \
-    --version=$LUXDEPS_VERSION \
-    --remote=mycenter \
-    --remote=mylocal \
-    --build=missing
+  EMBREE_PROFILE=$CLANG_PROFILE_ARM64
 fi
+
+cd $WORKSPACE/conan-local-recipes/recipes/embree/all
+conan create . \
+  --profile:all=$EMBREE_PROFILE \
+  --version=4.4.1 \
+  --build=embree* \
+  --build=missing \
+  --remote=conancenter
+cd $WORKSPACE
+
+conan create $WORKSPACE \
+  --profile:all=$CONAN_PROFILE \
+  --version=$LUXDEPS_VERSION \
+  --remote=mycenter \
+  --remote=mylocal \
+  --remote=conancenter \
+  --build=missing \
+  --build=!embree*
 echo "::endgroup::"
 
 # 8. Save result
@@ -240,26 +200,14 @@ if [[ -n "$CI" ]]; then
 fi
 conan remove -c -vverbose "*/*#!latest"  # Keep only latest version of each package
 # Save only dependencies of current target (otherwise cache gets bloated)
-if [[ "$RUNNER_OS" == "Windows" && "$RUNNER_ARCH" == "ARM64" ]];then
-  conan graph info \
-    --requires=luxcoredeps/$LUXDEPS_VERSION@luxcore/luxcore \
-    --requires=luxcoreconf/$LUXDEPS_VERSION@luxcore/luxcore \
-    --format=json \
-    --remote=mycenter \
-    --remote=mylocal \
-    --remote=mylocal-arm64 \
-    --profile:all=$CONAN_PROFILE \
-    > graph.json
-else
-  conan graph info \
-    --requires=luxcoredeps/$LUXDEPS_VERSION@luxcore/luxcore \
-    --requires=luxcoreconf/$LUXDEPS_VERSION@luxcore/luxcore \
-    --format=json \
-    --remote=mycenter \
-    --remote=mylocal \
-    --profile:all=$CONAN_PROFILE \
-    > graph.json
-fi
+conan graph info \
+  --requires=luxcoredeps/$LUXDEPS_VERSION@luxcore/luxcore \
+  --requires=luxcoreconf/$LUXDEPS_VERSION@luxcore/luxcore \
+  --format=json \
+  --remote=mycenter \
+  --remote=mylocal \
+  --profile:all=$CONAN_PROFILE \
+  > graph.json
 conan list --graph=graph.json --format=json --graph-binaries=Cache > list.json
 conan cache save -vverbose --file=${cache_dir}/conan-cache-save.tgz --list=list.json
 # Save build info
